+++ ts_src/types.ts
@@ -62,9 +62,9 @@ export const Network = typeforce.compile({
     public: typeforce.UInt32,
     private: typeforce.UInt32,
   },
-  pubKeyHash: typeforce.UInt8,
-  scriptHash: typeforce.UInt8,
-  wif: typeforce.UInt8,
+  pubKeyHash: typeforce.UInt16,
+  scriptHash: typeforce.UInt16,
+  wif: typeforce.UInt16,
 });
 
 export interface XOnlyPointAddTweakResult {
